Moved useResizeObserver to utils
