+++ packages/@react-spectrum/breadcrumbs/src/Breadcrumbs.tsx
@@ -12,11 +12,11 @@
 
 import {ActionButton} from '@react-spectrum/button';
 import {BreadcrumbItem} from './BreadcrumbItem';
-import {classNames, useDOMRef, useStyleProps} from '@react-spectrum/utils';
+import {classNames, useDOMRef, useStyleProps, useResizeObserver} from '@react-spectrum/utils';
 import {DOMRef} from '@react-types/shared';
 import FolderBreadcrumb from '@spectrum-icons/ui/FolderBreadcrumb';
 import {Menu, MenuTrigger} from '@react-spectrum/menu';
-import React, {Key, ReactElement, RefObject, useCallback, useEffect, useLayoutEffect, useRef, useState} from 'react';
+import React, {Key, ReactElement, useCallback,  useLayoutEffect, useRef, useState} from 'react';
 import {SpectrumBreadcrumbsProps} from '@react-types/breadcrumbs';
 import styles from '@adobe/spectrum-css-temp/components/breadcrumb/vars.css';
 import {useBreadcrumbs} from '@react-aria/breadcrumbs';
@@ -25,44 +25,6 @@ import {useProviderProps} from '@react-spectrum/provider';
 const MIN_VISIBLE_ITEMS = 1;
 const MAX_VISIBLE_ITEMS = 4;
 
-function hasResizeObserver() {
-  return typeof window.ResizeObserver !== 'undefined';
-}
-
-type useResizeObserverOptionsType<T> = {
-  ref: RefObject<T>,
-  onResize: () => void
-}
-
-function useResizeObserver<T extends HTMLElement>(options: useResizeObserverOptionsType<T>) {
-  const {ref, onResize} = options;
-
-  useEffect(() => {
-    if (!ref) { return; }
-
-    if (!hasResizeObserver()) {
-      window.addEventListener('resize', onResize, false);
-      return () => {
-        window.removeEventListener('resize', onResize, false);
-      };
-    } else {
-
-      const resizeObserverInstance = new ResizeObserver((entries) => {
-        if (!entries.length) {
-          return;
-        }
-
-        onResize();
-      });
-      resizeObserverInstance.observe(ref.current);
-
-      return () => {
-        resizeObserverInstance.unobserve(ref.current);
-      };
-    }
-
-  }, [onResize, ref]);
-}
 
 
 function Breadcrumbs<T>(props: SpectrumBreadcrumbsProps<T>, ref: DOMRef) {
